Remove texture normalization from vertex buffer conversion. Instead, apply the texture scale factor in our vertex shader HLSL This removed yet another reason for buffer patching, simplifying code more and speeding up rendering a little.
The Ripple XDK sample regressed because of this (or an earlier commit?), which might (or might not) be related to vertex explosions seen in some games.

(That, or it has something to do with the use of non-standard registers for passing in texture coordinates - in any case, a fix for this will probably improve a few games as well).